Pequeños cambios

+++ GestorExamenes/application/controllers/Auth.php
@@ -32,11 +32,12 @@ class Auth extends CI_Controller
      */
     public function index()
     {
-/*        if (file_exists("C:\\xampp\htdocs\\GestorExamenes\\application\controllers\\Instalacion.php"))
+        if (file_exists("C:\\xampp\htdocs\\GestorExamenes\\application\controllers\\Instalacion.php"))
             unlink("C:\\xampp\htdocs\\GestorExamenes\\application\controllers\\Instalacion.php");
         if (file_exists("C:\\xampp\htdocs\\GestorExamenes\\application\views\\instalacion.php"))
-            unlink("C:\\xampp\htdocs\\GestorExamenes\\application\views\\instalacion.php"); la linea se descomentará cuando se presente el proyecto, para evitar
-el borrado accidental de los metodos de instalacion    */
+            unlink("C:\\xampp\htdocs\\GestorExamenes\\application\views\\instalacion.php");
+        /*la linea se descomentará cuando se presente el proyecto, para evitar
+        el borrado accidental de los metodos de instalacion    */
     }
 
 
@@ -86,6 +87,7 @@ el borrado accidental de los metodos de instalacion    */
         $data['google_login_url'] = $this->google->get_login_url();
         redirect(base_url());
     }
+
     public function comprobacion()
     {
         $data['google_login_url'] = $this->google->get_login_url();
